Prevent error when wikidata has a claim with a no value snak
When a wikidata page had a claim with with snaktype = "novalue" the
reduce_claims function would not set the snak variable, triggering an
UnboundLocalError exception. This change prevents this by skipping to
the next snak when a novalue snak is encountered.

We also use the ``.items()`` method on the claims dictionary to avoid an
unnecessary lookup.